Volley has been completely removed from the project! Fucking finally! Because of that a lot of java classes had to be converted to Kotlin which can cause endless bugs.

+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/model/orm/Loadable.java
@@ -296,7 +296,7 @@ public class Loadable
     }
 
     public String desktopUrl() {
-        return site.resolvable().desktopUrl(this, no);
+        return site.resolvable().desktopUrl(this, (long) no);
     }
 
     public boolean isSuitableForPrefetch() {
